requirements/product.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
coverage==4.5.2
decorator==5.1.1
Django==3.0.3
djangorestframework==3.11.0
django-sslserver==0.22
executing==1.2.0
flake8==4.0.1
google-api-core==2.11.0
google-api-python-client==2.88.0
google-auth==2.19.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
jedi==0.18.2
line-profiler==3.5.1
matplotlib-inline==0.1.6
mccabe==0.6.1
mysqlclient==2.1.1
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
protobuf==4.23.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.15.1
pyparsing==3.0.9
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
toml==0.10.2
traitlets==5.9.0
typing_extensions==4.6.3
uritemplate==4.1.1
urllib3==1.26.16
wcwidth==0.2.6
